+++ IsekaiMod/Content/Classes/IsekaiProtagonist/Archetypes/Villain.cs
@@ -48,11 +48,11 @@ namespace IsekaiMod.Content.Classes.IsekaiProtagonist.Archetypes
                 bp.RemoveFeatures = new LevelEntry[] {
                     Helpers.LevelEntry(1, IsekaiProtagonistProficiencies, SneakAttack),
                     Helpers.LevelEntry(3, SpecialPowerSelection),
-                    Helpers.LevelEntry(5, SneakAttack, OverpoweredAbilitySelection2),
+                    Helpers.LevelEntry(5, SneakAttack),
                     Helpers.LevelEntry(9, SpecialPowerSelection, SneakAttack, FriendlyAuraFeature),
                     Helpers.LevelEntry(10, OverpoweredAbilitySelection2),
                     Helpers.LevelEntry(13, SneakAttack),
-                    Helpers.LevelEntry(15, OverpoweredAbilitySelection2, IsekaiQuickFooted),
+                    Helpers.LevelEntry(15, IsekaiQuickFooted),
                     Helpers.LevelEntry(17, SneakAttack),
                     Helpers.LevelEntry(20, SecondReincarnation),
                 };