+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Armor Saint` Special power.
 - Added `Playful Darkness` to `Summon Calamity` Overpowered ability.
 - Added bard, ranger, paladin, and inquisitor spells into the `Isekai Protagonist` spellbook.
+- Added `Mythic Special Power` to mythic ability selection.
 - Added `Signature Ability` to `Isekai Protagonist` progression. This can be chosen instead of `Signature Attack`.
 - Added `Second Reincarnation` to `Isekai Protagonist` progression. This feature replaces `True Main Character`.
 - Added `Summon Harem` ability to `Isekai Protagonist` progression. This ability replaces `Harem Magnet`.
 - Added `Channel Energy` selection to `God Emperor` progression.
 - Added `Divine Array` feature to `God Emperor` progression.
+- Added `Celestial Realm` feature to `God Emperor` progression.
 ### Changed
 - Reworked `Isekai Protagonist` class and all archetypes.
 - Buffed `Nascent Apotheosis` to also give spell penetration and spell resistance.